math.h: move and add some floating point functions
Moved exp.c > double.c: ceil, copysign, fabs, floor, fmod,  modf, modff, round, signbit, trunc
Added to double.c and math.h: signbit(), copysign()

JIRA: PD-189
