+++ math/Makefile
@@ -4,4 +4,4 @@
 # Copyright 2012-2015, 2020 Phoenix Systems
 #
 
-OBJS += $(addprefix $(PREFIX_O)math/, trig.o power.o exp.o common.o hyper.o)
+OBJS += $(addprefix $(PREFIX_O)math/, double.o trig.o power.o exp.o common.o hyper.o)
