+++ math/exp.c
@@ -3,7 +3,7 @@
  *
  * libphoenix
  *
- * exp, frexp, ldexp, log, log10, modf, ceil, floor, fmod, fabs
+ * exp, frexp, ldexp, log, log10
  *
  * Copyright 2017 Phoenix Systems
  * Author: Aleksander Kaminski
@@ -120,50 +120,6 @@ double log10(double x)
 }
 
 
-double modf(double x, double* intpart)
-{
-	conv_t *conv = (conv_t *)&x;
-	double tmp = x;
-	int exp = conv->i.exponent - 1023;
-	uint64_t m, mask = 0xfffffffffffffLL;
-
-	if (exp > 52) {
-		*intpart = x;
-		return conv->i.sign? -0.0 : 0.0;
-	}
-	else if (exp < 0) {
-		*intpart = conv->i.sign? -0.0 : 0.0;
-		return x;
-	}
-
-	conv->i.mantisa = conv->i.mantisa & ~(mask >> exp);
-	*intpart = x;
-	x = tmp;
-
-	m = conv->i.mantisa;
-	m &= mask >> exp;
-
-	if (m == 0)
-		return 0.0;
-
-	conv->i.mantisa = m & mask;
-	normalizeSub(&x, &exp);
-
-	conv->i.exponent = exp + 1023;
-
-	return x;
-}
-
-float modff(float x, float* intpart)
-{
-	double ret, tmp;
-
-	ret = modf(x, &tmp);
-	*intpart = tmp;
-
-	return ret;
-}
-
 /* Uses quick powering and Maclaurin series to calculate value of e^x */
 double exp(double x)
 {
@@ -193,82 +149,4 @@ double exp(double x)
 	}
 
 	return res * resi;
-}
-
-
-double ceil(double x)
-{
-	double ipart, fpart;
-
-	fpart = modf(x, &ipart);
-
-	if (x > 0 && fpart + x != x)
-		ipart += 1.0;
-
-	return ipart;
-}
-
-
-double floor(double x)
-{
-	double ipart, fpart;
-
-	fpart = modf(x, &ipart);
-
-	if (x < 0 && fpart + x != x)
-		ipart -= 1.0;
-
-	return ipart;
-}
-
-
-double fmod(double number, double denom)
-{
-	double result, tquot;
-
-	if (denom == 0.0 || denom == -0.0)
-		return NAN;
-
-	if (denom == INFINITY || denom == -INFINITY)
-		return number;
-
-	modf(number / denom, &tquot);
-	result = tquot * denom;
-
-	return number - result;
-}
-
-
-double round(double x)
-{
-	double ret, frac;
-
-	frac = modf(x, &ret);
-
-	if (frac >= 0.5)
-		ret += 1.0;
-	else if (frac <= -0.5)
-		ret -= 1.0;
-
-	return ret;
-}
-
-
-double trunc(double x)
-{
-	double ret;
-
-	modf(x, &ret);
-
-	return ret;
-}
-
-
-double fabs(double x)
-{
-	conv_t *conv = (conv_t *)&x;
-
-	conv->i.sign = 0;
-
-	return x;
 }
